Fix warnings reported by clang(-tidy) 14

+++ examples/1d_stencil/1d_stencil_4_parallel.cpp
@@ -58,18 +58,18 @@ struct partition_data
       : data_(new double[size])
       , size_(size)
     {
-        double base_value = double(initial_value * size);
+        double base_value = initial_value * static_cast<double>(size);
         for (std::size_t i = 0; i != size; ++i)
-            data_[i] = base_value + double(i);
+            data_[static_cast<std::ptrdiff_t>(i)] = base_value + double(i);
     }
 
     double& operator[](std::size_t idx)
     {
-        return data_[idx];
+        return data_[static_cast<std::ptrdiff_t>(idx)];
     }
     double operator[](std::size_t idx) const
     {
-        return data_[idx];
+        return data_[static_cast<std::ptrdiff_t>(idx)];
     }
 
     std::size_t size() const
