+++ libs/pika/resource_partitioner/examples/oversubscribing_resource_partitioner.cpp
@@ -55,7 +55,8 @@ void do_stuff(std::size_t n, bool printout)
         std::cout << "[do stuff] " << n << "\n";
     for (std::size_t i(0); i < n; ++i)
     {
-        double f = std::sin(2 * M_PI * i / n);
+        double f = std::sin(
+            2 * M_PI * static_cast<double>(i) / static_cast<double>(n));
         if (printout)
             std::cout << "sin(" << i << ") = " << f << ", ";
     }
